requirements.txt
Flask>=2.2.3
Jinja2>=3.1.2
netmiko>=4.2.0
pandas>=1.4.2
vsdx>=0.5.13
pytest>=7.2.2
ttp>=0.9.2
pywebview>=4.0.2
loguru>=0.7.0
python-dotenv>=1.0.0